\name{libpressio-package}
\alias{libpressio-package}
\alias{libpressio}
\docType{package}
\title{
  A Generic Abstraction for Compression
}
\description{
  This library provides a R wrapper for the libpressio library.
  More information on libpressio can be found at https://github.com/CODARCode/libpressio
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Robert Underwood

Maintainer: Robert Underwood <robertu@g.clemson.edu>
}
\references{
}
\keyword{ libpressio, compression }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
